nexusflow/services/router/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexusflow-router"
version = "0.1.0"
description = "NexusFlow policy-based routing engine for insights and threads"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "nats-py>=2.7",
    "opentelemetry-sdk>=1.24",
    "nexusflow-core-models",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
